// packages/api/server/models/majicModel/findOwners/RawOwner.d.ts
export type RawOwner = {
    ccodrotxt: string, // Code du droit réel ou particulier
    ccodemtxt: string, // Code du démembrement / indivision
    dqualp: string, // Qualité du propriétaire
    dnomus: string, // Nom du propriétaire
    dprnus: string, // Prénom du propriétaire
    catpro2txt: string, // Classification de personne morale niveau 2
    dlign3: string, // 3ème ligne d'adresse
    dlign4: string, // 4ème ligne d'adresse
    dlign5: string, // 5ème ligne d'adresse
    dlign6: string, // 6ème ligne d'adresse
};